Move JSONFunctionTest from android to data test package so we can run it on iOS
